rearchitecturation in prevision to add selector type of schema
